+++ src/1050.c
@@ -7,3 +7,63 @@
 #pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80025E30.s")
 
 #pragma GLOBAL_ASM("asm/nonmatchings/1050/__dummy.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80025F40.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80026C88.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80026CB8.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80026E40.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80026FC8.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80027148.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80027250.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80027408.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80027500.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80027660.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_800276A4.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80029C78.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_80029F60.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002A2C0.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002A8D8.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002ABA8.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002B24C.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002BCC8.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002BF54.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C2F4.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C4C4.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C500.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C568.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C5E4.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C69C.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C6D8.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C700.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C790.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C7B8.s")
+
+#pragma GLOBAL_ASM("asm/nonmatchings/1050/func_8002C918.s")
